# docs/swagger.yaml
definitions:
  http.Avatar_obj:
    properties:
      media_id:
        type: string
      url:
        type: string
    type: object
...
  http.address:
    properties:
      city:
        type: string
      district:
        type: string
      street:
        type: string
    type: object
...
  http.categoryResp:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  http.getAllProductsResp:
    properties:
      items:
        items:
          $ref: '#/definitions/http.getAllProductsRespItem'
        type: array
      meta:
        $ref: '#/definitions/internal_shop_delivery_http.listMetaResponse'
    type: object
  http.getAllProductsRespItem:
    properties:
      categories:
        items:
          $ref: '#/definitions/http.categoryResp'
        type: array
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/http.mediaResp'
        type: array
      inventory:
        $ref: '#/definitions/http.inventoryResp'
      name:
        type: string
      shop:
        $ref: '#/definitions/http.shopResp'
    type: object
...
  http.getDetailResp:
    properties:
      address:
        $ref: '#/definitions/http.address'
      avatar_obj:
        $ref: '#/definitions/http.Avatar_obj'
      avg_rate:
        type: number
      created_at:
        type: string
      followers:
        items:
          type: string
        type: array
      id:
        type: string
      is_verified:
        type: boolean
      name:
        type: string
      phone:
        type: string
    type: object
...
  http.getShopResp:
    properties:
      items:
        items:
          $ref: '#/definitions/http.getShopRespItem'
        type: array
      meta:
        $ref: '#/definitions/internal_shop_delivery_http.listMetaResponse'
    type: object
  http.getShopRespItem:
    properties:
      address:
        $ref: '#/definitions/http.address'
      avatar_obj:
        $ref: '#/definitions/http.Avatar_obj'
      avg_rate:
        type: number
      created_at:
        type: string
      followers:
        items:
          type: string
        type: array
      id:
        type: string
      is_verified:
        type: boolean
      name:
        type: string
      phone:
        type: string
      user_id:
        type: string
    type: object
...
  http.inventoryResp:
    properties:
      id:
        type: string
    type: object
...
  http.mediaResp:
    properties:
      media_id:
        type: string
      url:
        type: string
    type: object
...
  http.registerResponse:
    properties:
      address:
        $ref: '#/definitions/http.address'
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  http.shopResp:
    properties:
      city:
        type: string
      district:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      street:
        type: string
    type: object
...
  http.updateResp:
    properties:
      address:
        $ref: '#/definitions/http.address'
      avg_rate:
        type: number
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
  internal_shop_delivery_http.listMetaResponse:
    properties:
      count:
        type: integer
      current_page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  response.Resp:
    properties:
      data: {}
      error_code:
        type: integer
      errors: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/shops:
    delete:
      consumes:
      - application/json
      description: Delete a shop
      parameters:
      - default: '"*"'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NjAxMTk2NjgsImlhdCI6MTcyODU4MzY2OCwic3ViIjoiNjcwNzgyNWQ0NTgwNGNhYWY4MzE2OTU3Iiwic2Vzc2lvbl9pZCI6InpnSFJMd1NmTnNQVnk2d2g3M0ZLVmpqZXV6T1ZnWGZSMjdRYVd1eGtsdzQ9IiwidHlwZSI6IiIsInJlZnJlc2giOmZhbHNlfQ.Pti0gJ5fO4WjGTsxShGv90pr0E_0jMJdWFEUJYKG4VU
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 6707825d45804caaf8316957
        description: User ID
        in: header
        name: x-client-id
        required: true
        type: string
      - default: zgHRLwSfNsPVy6wh73FKVjjeuzOVgXfR27QaWuxklw4=
        description: Session ID
        in: header
        name: session-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Delete a shop
      tags:
      - Shop
    get:
      consumes:
      - application/json
      description: Get shop by Search, IDs,...
      parameters:
      - default: '"*"'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NjAxMTk2NjgsImlhdCI6MTcyODU4MzY2OCwic3ViIjoiNjcwNzgyNWQ0NTgwNGNhYWY4MzE2OTU3Iiwic2Vzc2lvbl9pZCI6InpnSFJMd1NmTnNQVnk2d2g3M0ZLVmpqZXV6T1ZnWGZSMjdRYVd1eGtsdzQ9IiwidHlwZSI6IiIsInJlZnJlc2giOmZhbHNlfQ.Pti0gJ5fO4WjGTsxShGv90pr0E_0jMJdWFEUJYKG4VU
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 6707825d45804caaf8316957
        description: User ID
        in: header
        name: x-client-id
        required: true
        type: string
      - default: zgHRLwSfNsPVy6wh73FKVjjeuzOVgXfR27QaWuxklw4=
        description: Session ID
        in: header
        name: session-id
        required: true
        type: string
      - description: Search
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: IDs
        in: query
        items:
          type: string
        name: ids
        type: array
      - description: Is Verified
        in: query
        name: is_verified
        type: boolean
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.getShopResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Get shop
      tags:
      - Shop
    patch:
      consumes:
      - application/json
      description: Update a shop
      parameters:
      - default: '"*"'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NjAxMTk2NjgsImlhdCI6MTcyODU4MzY2OCwic3ViIjoiNjcwNzgyNWQ0NTgwNGNhYWY4MzE2OTU3Iiwic2Vzc2lvbl9pZCI6InpnSFJMd1NmTnNQVnk2d2g3M0ZLVmpqZXV6T1ZnWGZSMjdRYVd1eGtsdzQ9IiwidHlwZSI6IiIsInJlZnJlc2giOmZhbHNlfQ.Pti0gJ5fO4WjGTsxShGv90pr0E_0jMJdWFEUJYKG4VU
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 6707825d45804caaf8316957
        description: User ID
        in: header
        name: x-client-id
        required: true
        type: string
      - default: zgHRLwSfNsPVy6wh73FKVjjeuzOVgXfR27QaWuxklw4=
        description: Session ID
        in: header
        name: session-id
        required: true
        type: string
      - description: IDs
        in: body
        name: ids
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.updateResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Update a shop
      tags:
      - Shop
    post:
      consumes:
      - application/json
      description: Create a shop
      parameters:
      - default: '"*"'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NjAxMTk2NjgsImlhdCI6MTcyODU4MzY2OCwic3ViIjoiNjcwNzgyNWQ0NTgwNGNhYWY4MzE2OTU3Iiwic2Vzc2lvbl9pZCI6InpnSFJMd1NmTnNQVnk2d2g3M0ZLVmpqZXV6T1ZnWGZSMjdRYVd1eGtsdzQ9IiwidHlwZSI6IiIsInJlZnJlc2giOmZhbHNlfQ.Pti0gJ5fO4WjGTsxShGv90pr0E_0jMJdWFEUJYKG4VU
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 6707825d45804caaf8316957
        description: User ID
        in: header
        name: x-client-id
        required: true
        type: string
      - default: zgHRLwSfNsPVy6wh73FKVjjeuzOVgXfR27QaWuxklw4=
        description: Session ID
        in: header
        name: session-id
        required: true
        type: string
      - description: Name
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: Phone
        in: body
        name: phone
        required: true
        schema:
          type: string
      - description: City
        in: body
        name: city
        required: true
        schema:
          type: string
      - description: Street
        in: body
        name: street
        required: true
        schema:
          type: string
      - description: District
        in: body
        name: district
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.registerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Create a shop
      tags:
      - Shop
  /api/v1/shops/{id}:
    get:
      consumes:
      - application/json
      description: Get shop detail by ID
      parameters:
      - default: '"*"'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NjAxMTk2NjgsImlhdCI6MTcyODU4MzY2OCwic3ViIjoiNjcwNzgyNWQ0NTgwNGNhYWY4MzE2OTU3Iiwic2Vzc2lvbl9pZCI6InpnSFJMd1NmTnNQVnk2d2g3M0ZLVmpqZXV6T1ZnWGZSMjdRYVd1eGtsdzQ9IiwidHlwZSI6IiIsInJlZnJlc2giOmZhbHNlfQ.Pti0gJ5fO4WjGTsxShGv90pr0E_0jMJdWFEUJYKG4VU
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 6707825d45804caaf8316957
        description: User ID
        in: header
        name: x-client-id
        required: true
        type: string
      - default: zgHRLwSfNsPVy6wh73FKVjjeuzOVgXfR27QaWuxklw4=
        description: Session ID
        in: header
        name: session-id
        required: true
        type: string
      - description: Shop ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.getDetailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Get shop detail
      tags:
      - Shop
...
  /api/v1/shops/get-shop-id-by-user-id/{id}:
    get:
      consumes:
      - application/json
      description: Get shop ID by user ID
      parameters:
      - default: '"*"'
        description: Access-Control-Allow-Origin
        in: header
        name: Access-Control-Allow-Origin
        type: string
      - default: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NjAxMTk2NjgsImlhdCI6MTcyODU4MzY2OCwic3ViIjoiNjcwNzgyNWQ0NTgwNGNhYWY4MzE2OTU3Iiwic2Vzc2lvbl9pZCI6InpnSFJMd1NmTnNQVnk2d2g3M0ZLVmpqZXV6T1ZnWGZSMjdRYVd1eGtsdzQ9IiwidHlwZSI6IiIsInJlZnJlc2giOmZhbHNlfQ.Pti0gJ5fO4WjGTsxShGv90pr0E_0jMJdWFEUJYKG4VU
        description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 6707825d45804caaf8316957
        description: User ID
        in: header
        name: x-client-id
        required: true
        type: string
      - default: zgHRLwSfNsPVy6wh73FKVjjeuzOVgXfR27QaWuxklw4=
        description: Session ID
        in: header
        name: session-id
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Shop ID
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Get shop ID by user ID
      tags:
      - Shop
...
  /api/v1/shops/products/all:
    get:
      consumes:
      - application/json
      description: Get all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.getAllProductsResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Resp'
      summary: Get all products
      tags:
      - Shop
...
swagger: "2.0"